requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.0
httpx~=0.27.0
cryptography>=39.0.1
cachetools>=5.3.3
APScheduler>=3.10.4
flask==3.0.2
waitress==3.0.0
pycoingecko==3.1.0
pandas==2.1.4
matplotlib==3.8.2
python-binance==1.0.19
mplfinance==0.12.10b0
beautifulsoup4==4.12.2
requests==2.31.0    